refactor(Line): PointJudge 메서드 Line 내부로 이동
